fix get brand id from context as int not int64

+++ service/email_otp_service.go
@@ -106,7 +106,7 @@ func (service *EmailOtpService) sendEmail(c *gin.Context, otp string) error {
 			Otp:      otp,
 			Provider: utilities.MailGunName,
 			DateTime: time.Now().Format(time.DateTime),
-			BrandId:  c.GetInt64("_brand"),
+			BrandId:  int64(c.GetInt("_brand")),
 		},
 	}
 	if err := model.LogOtpEvent(event); err != nil {
